Add files via upload
In order to open these projects, you will have to download them by going to code, zip and download as zip file, then on your computer go to where the file was downloaded and extract the file, afterwards go to this website: "https://www.online-python.com/" and remove the code present there and open anyone of the files found in the folder, feel free to modify as you wish! Also feel free to comment anything to help or anything that you might want for me to personalize for you!